Added Separation For Code Readability
