+++ lib/address_check_options.dart
@@ -1,0 +1,22 @@
+part of data_connection_checker;
+
+/// This class should be pretty self-explanatory.
+/// If [AddressCheckOptions.port]
+/// or [AddressCheckOptions.timeout] are not specified, they both
+/// default to [DataConnectionChecker.DEFAULT_PORT]
+/// and [DataConnectionChecker.DEFAULT_TIMEOUT]
+/// Also... yeah, I'm not great at naming things.
+class AddressCheckOptions {
+  final InternetAddress address;
+  final int port;
+  final Duration timeout;
+
+  AddressCheckOptions(
+    this.address, {
+    this.port = DataConnectionChecker.DEFAULT_PORT,
+    this.timeout = DataConnectionChecker.DEFAULT_TIMEOUT,
+  });
+
+  @override
+  String toString() => 'AddressCheckOptions($address, $port, $timeout)';
+}
