+++ lib/status_enum.dart
@@ -1,0 +1,8 @@
+part of data_connection_checker;
+
+/// Represents the status of the data connection.
+/// Returned by [DataConnectionChecker.connectionStatus]
+enum DataConnectionStatus {
+  disconnected,
+  connected,
+}
